--- lib/constants.ts
@@ -1,5 +1,0 @@
-export const LOGO_URL =
-  'https://images.ctfassets.net/5ct6f2q9wwt3/2dQg3rHxURIhQCjK3Oxds2/10460c53028ab886bd8a1c925459cf53/FFLTrainer-logo-V1.png'
-export const RANDOM_IMG_URL = '//source.unsplash.com/random/300×300/?sport'
-export const GITHUB_URL = 'https://github.com/james-langridge'
-export const FACEBOOK_URL = 'https://www.facebook.com/fitforlifetrainer.co.uk'
